+++ controller/callbackcontroller.php
@@ -168,6 +168,23 @@ class CallbackController extends Controller {
             return new JSONResponse(["message" => $this->trans->t("Invalid request")], Http::STATUS_BAD_REQUEST);
         }
 
+        if (!empty($this->config->GetDocumentServerSecret())) {
+            $header = \OC::$server->getRequest()->getHeader("Authorization");
+            if (empty($header)) {
+                $this->logger->info("Download without jwt", array("app" => $this->appName));
+                return new JSONResponse(["message" => $this->trans->t("Access deny")], Http::STATUS_FORBIDDEN);
+            }
+
+            $header = substr($header, strlen("Bearer "));
+
+            try {
+                $decodedHeader = \Firebase\JWT\JWT::decode($header, $this->config->GetDocumentServerSecret(), array("HS256"));
+            } catch (\UnexpectedValueException $e) {
+                $this->logger->info("Download with invalid jwt: " . $e->getMessage(), array("app" => $this->appName));
+                return new JSONResponse(["message" => $this->trans->t("Access deny")], Http::STATUS_FORBIDDEN);
+            }
+        }
+
         $fileId = $hashData->fileId;
         $ownerId = $hashData->ownerId;
 
@@ -185,7 +202,7 @@ class CallbackController extends Controller {
 
         try {
             return new DataDownloadResponse($file->getContent(), $file->getName(), $file->getMimeType());
-        } catch(\OCP\Files\NotPermittedException  $e) {
+        } catch (\OCP\Files\NotPermittedException  $e) {
             $this->logger->info("Download Not permitted: " . $fileId . " " . $e->getMessage(), array("app" => $this->appName));
             return new JSONResponse(["message" => $this->trans->t("Not permitted")], Http::STATUS_FORBIDDEN);
         }
@@ -216,6 +233,23 @@ class CallbackController extends Controller {
             return new JSONResponse(["message" => $this->trans->t("Invalid request")], Http::STATUS_BAD_REQUEST);
         }
 
+        if (!empty($this->config->GetDocumentServerSecret())) {
+            $header = \OC::$server->getRequest()->getHeader("Authorization");
+            if (empty($header)) {
+                $this->logger->info("Download empty without jwt", array("app" => $this->appName));
+                return new JSONResponse(["message" => $this->trans->t("Access deny")], Http::STATUS_FORBIDDEN);
+            }
+
+            $header = substr($header, strlen("Bearer "));
+
+            try {
+                $decodedHeader = \Firebase\JWT\JWT::decode($header, $this->config->GetDocumentServerSecret(), array("HS256"));
+            } catch (\UnexpectedValueException $e) {
+                $this->logger->info("Download empty with invalid jwt: " . $e->getMessage(), array("app" => $this->appName));
+                return new JSONResponse(["message" => $this->trans->t("Access deny")], Http::STATUS_FORBIDDEN);
+            }
+        }
+
         $templatePath = dirname(__DIR__) . DIRECTORY_SEPARATOR . "assets" . DIRECTORY_SEPARATOR . "en" . DIRECTORY_SEPARATOR . "new.docx";
 
         $template = file_get_contents($templatePath);
@@ -226,7 +260,7 @@ class CallbackController extends Controller {
 
         try {
             return new DataDownloadResponse($template, "new.docx", "application/vnd.openxmlformats-officedocument.wordprocessingml.document");
-        } catch(\OCP\Files\NotPermittedException  $e) {
+        } catch (\OCP\Files\NotPermittedException  $e) {
             $this->logger->info("Download Not permitted: " . $fileId . " " . $e->getMessage(), array("app" => $this->appName));
             return new JSONResponse(["message" => $this->trans->t("Not permitted")], Http::STATUS_FORBIDDEN);
         }
@@ -260,6 +294,30 @@ class CallbackController extends Controller {
             return new JSONResponse(["message" => $this->trans->t("Invalid request")], Http::STATUS_BAD_REQUEST);
         }
 
+        if (!empty($this->config->GetDocumentServerSecret())) {
+            $header = \OC::$server->getRequest()->getHeader("Authorization");
+            if (empty($header)) {
+                $this->logger->info("Track without jwt", array("app" => $this->appName));
+                return new JSONResponse(["message" => $this->trans->t("Access deny")], Http::STATUS_FORBIDDEN);
+            }
+
+            $header = substr($header, strlen("Bearer "));
+
+            try {
+                $decodedHeader = \Firebase\JWT\JWT::decode($header, $this->config->GetDocumentServerSecret(), array("HS256"));
+                $this->logger->debug("Track HEADER : " . json_encode($decodedHeader), array("app" => $this->appName));
+
+                $payload = $decodedHeader->payload;
+                $users = isset($payload->users) ? $payload->users : NULL;
+                $key = $payload->key;
+                $status = $payload->status;
+                $url = isset($payload->url) ? $payload->url : NULL;
+            } catch (\UnexpectedValueException $e) {
+                $this->logger->info("Track with invalid jwt: " . $e->getMessage(), array("app" => $this->appName));
+                return new JSONResponse(["message" => $this->trans->t("Access deny")], Http::STATUS_FORBIDDEN);
+            }
+        }
+
         $trackerStatus = $this->_trackerStatus[$status];
 
         $error = 1;
@@ -269,6 +327,9 @@ class CallbackController extends Controller {
 
                 $fileId = $hashData->fileId;
                 $ownerId = $hashData->ownerId;
+
+                \OC_Util::tearDownFS();
+                \OC_Util::setupFS($ownerId);
 
                 $files = $this->root->getUserFolder($ownerId)->getById($fileId);
                 if (empty($files)) {
@@ -308,7 +369,7 @@ class CallbackController extends Controller {
                         $from = $parsedUrl["scheme"] . "://" . $parsedUrl["host"] . (array_key_exists("port", $parsedUrl) ? (":" . $parsedUrl["port"]) : "") . "/";
                     }
 
-                    $this->logger->debug("Replace in track from " . $from . " to " . $this->config->GetDocumentServerInternalUrl(true));
+                    $this->logger->debug("Replace in track from " . $from . " to " . $this->config->GetDocumentServerInternalUrl(true), array("app" => $this->appName));
                     $url = str_replace($from, $this->config->GetDocumentServerInternalUrl(true), $url);
                 }